fully functioning selective repeat
To test the selective repeat and connecting functionality of the server and client.

You need to make sure you are using python 3.10 (to have match statement support), make sure you don't have python 3.12 specifically as there is a known bug regarding threading which won't allow the project to work. Use python3 --version to check your version

To test this,

Open one terminal in the UDP folder, and use the following command:

python3 server.py

Then open another terminal in the UDP folder and use the following command:

python3 client.py

If you look at the client.py file, the boot up sequence contains a send function for a dummy Hello Packet, which establishes the connection between the server and the client.

The server responds to the client's Hello Packet with a Hello Response Packet. And now at this point, both the server and the client have a conversation instance going for each other. From here on, it's all selective repeat.

For now for testing purposes, I made it so that the server and the client continuously respond to each other when either gets a hello or a hello response.

Right now, the sending and unpacking is slowed down so that you can see the process working properly.